Separate list. Initial array implementation
