+++ C_Tutor/C_Tutor.cbp
@@ -32,6 +32,14 @@
 			<Add option="-Wall" />
 			<Add option="-std=c11" />
 		</Compiler>
+		<Unit filename="array.c">
+			<Option compilerVar="CC" />
+		</Unit>
+		<Unit filename="array.h" />
+		<Unit filename="list.c">
+			<Option compilerVar="CC" />
+		</Unit>
+		<Unit filename="list.h" />
 		<Unit filename="main.c">
 			<Option compilerVar="CC" />
 		</Unit>
