+++ C_Tutor/C_Tutor.depend
@@ -11,3 +11,15 @@
 	<stdlib.h>
 	<stdint.h>
 
+1706814863 source:/home/yu3ga/Projects/Programming/ScriptsCode/C-main/C/C_Tutor/list.c
+	"list.h"
+
+1706815029 /home/yu3ga/Projects/Programming/ScriptsCode/C-main/C/C_Tutor/list.h
+	<stdio.h>
+	<stdlib.h>
+	<stdint.h>
+
+1706815186 source:/home/yu3ga/Projects/Programming/ScriptsCode/C-main/C/C_Tutor/main.c
+	<stdint.h>
+	"list.h"
+
